Renamed openapi spec
